requirements.txt
fastapi
pydantic
uvicorn
numpy
tensorflow
Pillow
python-multipart
httpx
h5py
mangum